// typescript-frontend/src/api/models/AvailabilityType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schedule API
 * API for scheduling
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const AvailabilityType = {
    Desired: 'DESIRED',
    Undesired: 'UNDESIRED',
    Unavailable: 'UNAVAILABLE'
} as const;
export type AvailabilityType = typeof AvailabilityType[keyof typeof AvailabilityType];


export function instanceOfAvailabilityType(value: any): boolean {
    for (const key in AvailabilityType) {
        if (Object.prototype.hasOwnProperty.call(AvailabilityType, key)) {
            if (AvailabilityType[key as keyof typeof AvailabilityType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function AvailabilityTypeFromJSON(json: any): AvailabilityType {
    return AvailabilityTypeFromJSONTyped(json, false);
}

export function AvailabilityTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): AvailabilityType {
    return json as AvailabilityType;
}

export function AvailabilityTypeToJSON(value?: AvailabilityType | null): any {
    return value as any;
}

export function AvailabilityTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): AvailabilityType {
    return value as AvailabilityType;
}