// typescript-frontend/src/api/models/EmployeeModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schedule API
 * API for scheduling
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EmployeeModel
 */
export interface EmployeeModel {
    /**
     * 
     * @type {string}
     * @memberof EmployeeModel
     */
    name: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof EmployeeModel
     */
    skillSet: Array<string>;
}

/**
 * Check if a given object implements the EmployeeModel interface.
 */
export function instanceOfEmployeeModel(value: object): value is EmployeeModel {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('skillSet' in value) || value['skillSet'] === undefined) return false;
    return true;
}

export function EmployeeModelFromJSON(json: any): EmployeeModel {
    return EmployeeModelFromJSONTyped(json, false);
}

export function EmployeeModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): EmployeeModel {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'skillSet': json['skill_set'],
    };
}

export function EmployeeModelToJSON(json: any): EmployeeModel {
    return EmployeeModelToJSONTyped(json, false);
}

export function EmployeeModelToJSONTyped(value?: EmployeeModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'skill_set': value['skillSet'],
    };
}